# magazine_api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AccountBase:
    properties:
      account_number:
        type: string
      bank_branch:
        type: string
      bank_name:
        type: string
      creator_id:
        type: string
      name:
        type: string
    type: object
...
  models.AddressBase:
    properties:
      contact_number:
        items:
          type: string
        type: array
      country:
        type: string
      creator_id:
        type: string
      district:
        type: string
      municipality:
        type: string
      state:
        type: string
      street:
        type: string
      ward:
        type: integer
    type: object
...
  models.ContactBase:
    properties:
      contact_number:
        type: string
      creator_id:
        type: string
      name:
        type: string
      relation:
        type: string
    type: object
...
  models.DocumentBase:
    properties:
      company_id:
        type: string
      creator_id:
        type: string
      document_type:
        type: string
      url:
        type: string
    type: object
...
  models.TailorRate:
    properties:
      product_id:
        type: string
      rate:
        type: number
    type: object
...
  models.User:
    properties:
      contact_number:
        type: string
      created_on:
        type: string
      deleted_on:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        items:
          type: string
        type: array
      updated_on:
        type: string
    type: object
  models.UserBase:
    properties:
      contact_number:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        items:
          type: string
        type: array
    type: object
...
  responses.EmployeeAll:
    properties:
      address:
        items:
          $ref: '#/definitions/models.AddressBase'
        type: array
      bank_account:
        items:
          $ref: '#/definitions/models.AccountBase'
        type: array
      company_id:
        type: string
      contact_number:
        type: string
      created_on:
        type: string
      deleted_on:
        type: string
      document:
        items:
          $ref: '#/definitions/models.DocumentBase'
        type: array
      email:
        type: string
      emergency_contact:
        items:
          $ref: '#/definitions/models.ContactBase'
        type: array
      employee_profile_id:
        type: string
      name:
        type: string
      picture:
        type: string
      role:
        items:
          type: string
        type: array
      salary:
        items:
          properties:
            amount:
              type: number
            effective_from:
              type: integer
            effective_to:
              type: integer
            salary_format:
              type: integer
            tailor_rate:
              items:
                $ref: '#/definitions/models.TailorRate'
              type: array
          type: object
        type: array
      user_id:
        type: string
    type: object
...
host: localhost:5005
info:
  contact:
    name: 21557870 - Suman Chhetri
  description: This is an API documentation of Magazine API
  title: Magazine API Documentation
  version: v1
paths:
...
  /employee/{id}:
    delete:
      description: Delete by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  type: string
              type: object
      summary: Soft Delete an user
      tags:
      - Employee
    patch:
      consumes:
      - application/json
      description: Updates user attribute
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserBase'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: Update an user
      tags:
      - Employee
  /employee/contact/{contact}:
    get:
      description: Gets One user by Contact Number
      parameters:
      - description: Contact Number
        in: path
        name: contact_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: Gets One User by Contact Number
      tags:
      - Employee
...
  /employee/email/{email}:
    get:
      description: Gets One user by email
      parameters:
      - description: Email
        format: email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
      summary: Gets One User by email
      tags:
      - Employee
  /employee/id/{id}:
    get:
      description: Get employee details by employee profile id.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/responses.EmployeeAll'
              type: object
      summary: Gets Employee By ID
      tags:
      - Employee
  /employee/profile/{id}:
    get:
      description: Get employee details by employee profile id.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/responses.EmployeeAll'
              type: object
      summary: Gets Employee profile by ID
      tags:
      - Employee
  /employee/type/{type}:
    get:
      description: List users by type
      parameters:
      - description: Type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
      summary: Lists users by type
      tags:
      - Employee
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"